Avoid #update on :set_nil parent resources being destroyed
When a top-level resource is being destroyed, there are options to cascade through that resource's associations to destroy the association or update the association so that it no longer points at this resource.  This goes awry, however, in the situation where there are two relationships between the classes, and one of them is set to :constraint => :destroy, while the other is set to :constraint => :set_nil.

So, for example, consider a data model where class A owns a set of class B instances, but declares one of them to be the best instance of class B:

  class A
    include DataMapper::Resource

    belongs_to :best_b,
               :model => B,
               :required => false
    has :n, bs, :constraint => :destroy
  end

  class B
    include DataMapper::Resource

    belongs_to :a
    has 0..1, as, :constraint => :set_nil
  end

In this situation, calling A.first.destroy() can result in an error of "A#update cannot be called on a dirty resource."  What happens is that the relationship constraints in class A cause the cascaded destruction of all the associated B instances.  The destruction of the "best_b" B instance causes the instance of A to be updated with best_b set to nil.  However, the instance of A is already being destroyed, so updating it is not allowed.

This patch will set an instance variable during the destruction of any resource.  When the time comes to :set_nil, the patch will check for the existence of that instance variable, and update the resource only if that instance variable is missing.  Finally, it removes the instance variable after the destruction is accomplished.